Add support for streaming the responses.
reorganize project a little
