+++ src/AIchat/utils/conversation.ts
@@ -1,0 +1,157 @@
+import * as fs from 'fs/promises';
+import { aiRoles } from '@AIchat/data/roles';
+import { promptModel } from './promptModel';
+import { saveChat } from './saveChat';
+import * as neovim from 'neovim';
+import * as bash from '@utils/bashHelper';
+import { geminiModels } from '../data/models';
+import { formatChatEntry } from './aiUtils';
+
+export async function converse(
+    chatFile: string,
+    temperature: number,
+    role: string,
+    model: string,
+    isChatLoaded = false,
+): Promise<void> {
+    try {
+        if (!isChatLoaded) {
+            await initializeUserPrompt(chatFile);
+        }
+
+        const socketPath = await generateSocketPath();
+
+        const tmuxCommand = `tmux split-window -v -p 90 -c "#{pane_current_path}" \; \
+        tmux send-keys -t :. 'sh -c "trap \\"exit 0\\" TERM; nvim --listen \\"${socketPath}\\" \\"${chatFile}\\"; tmux send-keys exit C-m"' C-m`
+
+        bash.execCommand(tmuxCommand);
+
+        await waitForSocket(socketPath);
+
+        const nvim = neovim.attach({ socket: socketPath });
+        const channelId = await nvim.channelId;
+
+        // create commands for saving and submitting
+        await nvim.command(`
+            function! SaveAndSubmit()
+                normal! o
+                write
+                call rpcnotify(${channelId}, 'prompt_action', 'submit_pressed')
+            endfunction
+            command! SubmitPrompt call SaveAndSubmit()
+        `);
+
+        await nvim.command(`nnoremap <CR> :SubmitPrompt<CR>`);
+
+        // open the file in Neovim
+        await nvim.command(`edit ${chatFile}`);
+        await updateNvimAndGoToLastLine(nvim);
+
+        await onHitEnterInNeovim(nvim, chatFile, model, temperature, role);
+
+        let intervalId: NodeJS.Timeout | undefined;
+        intervalId = setInterval(async () => {
+            try {
+                await fs.access(socketPath); // socket exists
+            } catch (error) { // socket does not exist
+                console.log('Chat Ended.');
+                clearInterval(intervalId);
+
+                const conversationHistory = await fs.readFile(chatFile, 'utf8');
+                const fullPrompt = conversationHistory + '\n';
+
+                await saveChat(chatFile, fullPrompt, temperature, role, model);
+
+                await fs.rm(chatFile);
+            }
+        }, 500);
+    } catch (error) {
+        console.error('Error in AI prompt workflow:', (error as Error).message);
+
+        throw error;
+    }
+}
+
+async function waitForSocket(socketPath: string, timeout = 5000) {
+    const start = Date.now();
+    while (Date.now() - start < timeout) {
+        try {
+            await fs.access(socketPath); // check if socket file exists
+            return true;
+        } catch (err) {
+            await new Promise(resolve => setTimeout(resolve, 200)); // retry every 200ms
+        }
+    }
+    return false;
+}
+
+async function generateSocketPath(): Promise<string> {
+    const randomString = Math.random().toString(36).substring(2, 15);
+    return `/tmp/nvim-${randomString}.sock`;
+}
+
+async function initializeUserPrompt(filePath: string): Promise<void> {
+    const initialContent = `# AI Chat History\n\nThis file contains the conversation history between the user and AI.\n\n---\n\n### USER (${new Date().toISOString()})\n\n`;
+    await fs.writeFile(filePath, initialContent, 'utf-8');
+}
+
+async function appendToChat(file: string, content: string): Promise<void> {
+    await fs.appendFile(file, content, 'utf-8');
+}
+
+async function updateNvimAndGoToLastLine(nvim: neovim.NeovimClient) {
+    await nvim.command('edit!');
+    await nvim.command('normal! G');
+    await nvim.command('normal! o');
+}
+
+async function onHitEnterInNeovim(nvim: neovim.NeovimClient, chatFile: string, model: string, temperature: number, role: string) {
+    return nvim.on('notification', async (method, args) =>
+    {
+        if (method === 'prompt_action' && args[0] === 'submit_pressed') {
+            const buffer = await nvim.buffer;
+            const lines = await buffer.lines;
+
+            const conversationHistory = lines.join('\n');
+            const fullPrompt = conversationHistory + '\n';
+
+            const aiEntryHeader = formatChatEntry(model, '', false);
+            await appendToChat(chatFile, aiEntryHeader);
+            await updateNvimAndGoToLastLine(nvim);
+
+            const response = await promptModel(model, fullPrompt, temperature, aiRoles[role]);
+
+            if (typeof response === 'string') {
+                await appendToChat(chatFile, response);
+                await appendToChat(chatFile, '\n');
+
+                await updateNvimAndGoToLastLine(nvim);
+            } else {
+                // streaming
+                let fullResponse = '';
+                try {
+                    for await (const chunk of response) {
+                        const text = chunk.text();
+                        fullResponse += text;
+
+                        await appendToChat(chatFile, text);
+                        await nvim.command('edit!');
+                    }
+
+                    // gemini adds an empty line at the end so exclude
+                    if (!geminiModels[model]) {
+                        await appendToChat(chatFile, '\n');
+                        await updateNvimAndGoToLastLine(nvim);
+                    }
+                } catch (error) {
+                    console.log(error);
+                }
+            }
+
+            await appendToChat(chatFile, '\n');
+            const userEntryHeader = formatChatEntry('USER', '', false);
+            await appendToChat(chatFile, userEntryHeader);
+            await updateNvimAndGoToLastLine(nvim);
+        }
+    });
+}
